+++ lib/video.js
@@ -1,0 +1,5 @@
+import duration from './video/duration'
+
+const video = { duration }
+
+export default video
