+++ lib/video/duration.js
